wayland window

+++ include/features/renderer/vulkan/vulkanDepthBuffer.h
@@ -16,7 +16,7 @@ namespace wne
 
         // width, height - resolution
         // isSampled - will depth be used with sampler in shaders
-        bool setup(uint16 width, uint16 height, uint64 sampledCountBit, bool isSampled = false);
+        bool setup(int width, int height, uint64 sampledCountBit, bool isSampled = false);
         void transitionToDefined();
 
         inline VulkanFormat getFormat()
@@ -45,7 +45,7 @@ namespace wne
         }
 
     protected:
-        uint16 width = 0, height = 0;
+        int width = 0, height = 0;
         VulkanFormat format;
 
         VulkanUtils *vulkanUtils = nullptr;
